Fix `rediscala-bench` build
Due to broken naming SBT interpreted the `rediscala` project (namely
`root`) and the `bench` package as same project. This bug caused several
compilation failures when trying to run the `bench` project as it missed
the required `rediscala` part.

Now we explicitly depend on `rediscala` in `bench` and fix the naming to
make it possible to run the benchmark tests.